src/backend/requirements.txt
# FastAPI & Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
aiosqlite>=0.19.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
cryptography>=42.0.0

# Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# PDF Processing
PyMuPDF>=1.23.0
pymupdf4llm>=0.0.5
pdfplumber>=0.10.0
Pillow>=10.2.0

# Web Scraping
playwright>=1.41.0

# LLM APIs
openai>=1.10.0
anthropic>=0.18.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.26.0
